+++ client/src/components/hue_page/index.js
@@ -1,0 +1,140 @@
+import React, { Component } from 'react';
+import { Container } from 'react-bulma-components';
+import Lights from '../lights';
+import axios from 'axios';
+
+class Hue_Page extends Component {
+  state = {
+    ip: [],
+    user: '',
+    lights: [],
+    lightId: [],
+    selectedLight: [],
+  }
+
+  componentDidMount() {
+    axios.post('/')
+      .then(res => {
+        let ip = res.data;
+        this.setState({ ip });
+      });
+
+    window.addEventListener("beforeunload", this.onUnload);
+    const stateObject = JSON.parse(localStorage.getItem("state"));
+    this.setState(stateObject);
+
+  }
+
+  onUnload = (event) => {
+    localStorage.setItem("state", JSON.stringify(this.state));
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener("beforeunload", this.onUnload)
+  }
+
+  connectionHandler = () => {
+    axios.post('/connect', {
+      host: this.state.ip
+    })
+      .then(res => {
+        console.log(res.data);
+        let user = res.data;
+        this.setState({ user })
+        this.findAllLights();
+      }).catch(function (error) {
+        if (error) alert('Make sure to hold bridge link when connecting!')
+      });
+  };
+
+  findAllLights = () => {
+    axios.post('/allLights', {
+      host: this.state.ip,
+      user: this.state.user
+    }).then(res => {
+      console.log(res.data)
+      let lights = res.data.lights.map(lights => lights.name);
+      let lightId = res.data.lights.map(lights => lights.id);
+      this.setState({ lights });
+      this.setState({ lightId })
+    });
+
+  };
+
+  handleChange = (event) => {
+    this.setState({ selectedLight: event.target.value })
+  };
+
+  lightOn = () => {
+    axios.post('/lights', {
+      host: this.state.ip,
+      username: this.state.user,
+      huestate: 'on',
+      light: this.state.selectedLight
+    })
+      .then(res => {
+        console.log(res);
+      });
+  };
+
+  lightOff = () => {
+    axios.post('/lights', {
+      host: this.state.ip,
+      username: this.state.user,
+      huestate: 'off',
+      light: this.state.selectedLight
+    })
+      .then(res => {
+        console.log(res);
+      });
+  };
+
+  criticalRoll = () => {
+    axios.post('/lights', {
+      host: this.state.ip,
+      username: this.state.user,
+      huestate: 'critical',
+      light: this.state.selectedLight
+    })
+      .then(res => {
+        console.log(res);
+      });
+  };
+
+  lightning = () => {
+    axios.post('/lights', {
+      host: this.state.ip,
+      username: this.state.user,
+      huestate: 'lightning',
+      light: this.state.selectedLight
+    })
+      .then(res => {
+        console.log(res);
+      });
+  };
+
+  render() {
+    return (
+      <Container id="hue-box">
+        <h1>Hue Lights</h1>
+        {this.state.ip.length > 0 ?
+          <div>
+            <h4>Select a Light:</h4>
+            <select onChange={this.handleChange} value={this.state.selectedLight}>
+              {this.state.lights.map((lights, index) => (
+                <option value={this.state.lightId[index]} key={this.state.lightId[index]}>{lights}</option>
+              ))}
+            </select>
+            <Lights
+              lightOn={this.lightOn}
+              lightOff={this.lightOff}
+              critical={this.criticalRoll}
+              lightning={this.lightning}
+              connection={this.connectionHandler}>
+            </Lights></div> : 'No Bridge Found'}
+      </Container>
+    );
+  }
+}
+
+export default Hue_Page;
