+++ client/src/components/lights/index.js
